Implement storage class
