+++ src/main/java/TaskList.java
@@ -1,0 +1,11 @@
+public class TaskList {
+    // Enum list to contain instructions to be entered by user to control system
+    public enum Instructions {
+        BYE, LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE
+    }
+
+    public void instructionList() {
+        
+    }
+
+}
