+++ src/main/java/Todo.java
@@ -1,0 +1,11 @@
+// To-do class returns result that is type [T]
+class Todo extends Task {
+    public Todo(String name) {
+        super(name);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
